Associate Categories and Items
Preliminary playing around... not really sure if I'm working in the right direction.
Also, I created a function to pull a list of all categories, but when I just printed it in the html page it wasn't the category names. Not sure why. Stopping here for now -- all my playing around is commented out.